+++ R/PSMA-package.R
@@ -2,6 +2,7 @@
 #' @importFrom stats complete.cases
 #'
 #' @importFrom hutils %notchin%
+#' @importFrom hutils selector
 #' @importFrom hutils trim_common_affixes
 #' @importFrom hutilscpp minmax
 #' @importFrom hutilscpp range_rcpp
@@ -9,6 +10,7 @@
 #' @importFrom stats setNames
 
 #' @importFrom utils head
+#' @importFrom utils hasName
 #' @importFrom utils packageVersion
 #' @importFrom utils download.file
 #' @importFrom utils unzip