+++ NEWS.md
@@ -1,12 +1,21 @@
-# PSMA 0.7.0
+## PSMA 0.8.0
+
+* New function:
+  * `avgLatLon_...` for average lat-lon pairs by postcode and street
+
+* Enhancements:
+  * `geocode` contemplates inexact addresses
+
+
+## PSMA 0.7.0
 * Update to Feb 2022
 * Provide `update_data` which accepts an unzipped folder from GNAF
 
-# PSMA 0.6.0
+## PSMA 0.6.0
 
 * Update to May 2021
 * Export `get_fst`.
 
-# PSMA 0.3.0
+## PSMA 0.3.0
 
 * Update to May 2018.